Implementacao do feedforward

+++ src/main/java/br/com/inteligenciaartificial/algoritmos/neural/TrainingDigit.java
@@ -1,15 +1,15 @@
 package br.com.inteligenciaartificial.algoritmos.neural;
 
 public class TrainingDigit extends Digit {
-	public final static int OUTPUT = 10;
-	public final double[] expectedOutput;
+    public final static int OUTPUT = 10;
+    public final int[] expectedOutput;
 
-	public TrainingDigit(final double[] pixels, final double[] expectedOutput) {
-		super(pixels);
-		if (expectedOutput == null || expectedOutput.length != OUTPUT) {
-			throw new IllegalArgumentException("Output of anyone digit must be " + OUTPUT);
-		}
-		this.expectedOutput = expectedOutput;
-	}
+    public TrainingDigit(final int[] pixels, final int[] expectedOutput) {
+        super(pixels);
+        if (expectedOutput == null || expectedOutput.length != OUTPUT) {
+            throw new IllegalArgumentException("Output of anyone digit must be " + OUTPUT);
+        }
+        this.expectedOutput = expectedOutput;
+    }
 
 }
